Downgrade MRIreco
After a patch change in the NFFTTools dependency, the MRIreco package is
not working properly.

In practice, in the KomaUI() it is not possible to reconstruct the image
from the "Raw Data" after clicking the "Recontruct!" button and the
process gets stuck indefinitely.

The problem has been documented in the MRIreco pull request
MagneticResonanceImaging/MRIReco.jl#95 and it is also solved, however it
has not been merged into the master branch yet.

The fast solution it is simply downgrade the MRIreco version from 0.5.0
to 0.4.2.